requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
iniconfig==1.1.1
Markdown==3.3.4
packaging==21.2
pluggy==1.0.0
psycopg2==2.9.1
py==1.11.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
python-environ==0.4.54
pytz==2021.3
sqlparse==0.4.2
termstyle==0.1.11
toml==0.10.2
whitenoise==5.3.0
